+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## Change Logs
 
 ### 1.1.0
+- Plugins:
+  - Add neo4j plugin.(#312)
 
+- Fixes:
+  - Fix unexpected 'No active span' IllegalStateError (#311)
+ 
 ### 1.0.1
 
 - Feature: